DockerAdvanced/SNM/old/snm_app/requirements.txt
flask
Flask-SQLAlchemy
mysql-connector-python